+++ gpAux/extensions/gps3ext/include/gpwriter.h
@@ -54,34 +54,4 @@ GPWriter *writer_init(const char *url_with_options);
 bool writer_transfer_data(GPWriter *writer, char *data_buf, int &data_len);
 bool writer_cleanup(GPWriter **writer);
 
-// Set schema to 'https' or 'http'
-inline string replaceSchemaFromURL(const string &url) {
-    size_t iend = url.find("://");
-    if (iend == string::npos) {
-        return url;
-    }
-    return "https" + url.substr(iend);
-}
-
-inline string getRegionFromURL(const string &url) {
-    size_t ibegin =
-        url.find("://s3") + strlen("://s3");  // index of character('.' or '-') after "3"
-    size_t iend = url.find(".amazonaws.com");
-
-    string region;
-    if (iend == string::npos) {
-        return region;
-    } else if (ibegin == iend) {  // "s3.amazonaws.com"
-        return "external-1";
-    } else {
-        // ibegin + 1 is the character after "s3." or "s3-"
-        // for instance: s3-us-west-2.amazonaws.com
-        region = url.substr(ibegin + 1, iend - (ibegin + 1));
-    }
-
-    if (region.compare("us-east-1") == 0) {
-        region = "external-1";
-    }
-    return region;
-}
 #endif /* INCLUDE_GPWRITER_H_ */
