+++ gpAux/extensions/gps3ext/include/s3key_writer.h
@@ -5,7 +5,7 @@
 #include "s3interface.h"
 #include "s3macros.h"
 #include "s3restful_service.h"
-#include "s3url_parser.h"
+#include "s3url.h"
 #include "writer.h"
 
 #include <string>
