+++ gpAux/extensions/gps3ext/test/s3url_test.cpp
@@ -1,4 +1,4 @@
-#include "s3url_parser.cpp"
+#include "s3url.cpp"
 #include "gtest/gtest.h"
 #include "s3macros.h"
